Policy Để Phân Quyền Chức Năng - Phần 2 112 Finish

+++ app/Http/Controllers/Admin/PostsController.php
@@ -12,21 +12,22 @@ class PostsController extends Controller
 {
     public function index()
     {
-        $userId=Auth::user()->id;
-        $lists = Post::orderBy('created_at','desc')
-        ->where('user_id',$userId)
-        ->get();
+        $userId = Auth::user()->id;
+        $lists = Post::orderBy('created_at', 'desc')
+            ->where('user_id', $userId)
+            ->get();
         return view('admin.posts.lists', compact('lists'));
     }
     public function add()
     {
         return view('admin.posts.add');
     }
-    public function postAdd(Request $request){
+    public function postAdd(Request $request)
+    {
         $request->validate(
             [
                 'title' => 'required',
-                'content'=>'required',
+                'content' => 'required',
             ],
             [
 
@@ -34,22 +35,24 @@ class PostsController extends Controller
                 'content.required' => 'Nội dung không được để trống',
             ]
         );
-    $post = new Post();
-    $post->title = $request->title;
-    $post->content = $request->content;
-    $post->user_id = Auth::user()->id;
-    $post->save();
+        $post = new Post();
+        $post->title = $request->title;
+        $post->content = $request->content;
+        $post->user_id = Auth::user()->id;
+        $post->save();
         return redirect()->route('admin.posts.index')->with('msg', 'Thêm bài viết thành công');
     }
-    public function edit(Post $post) {
-        $this->authorize('update',$post);
-        return view('admin.posts.edit',compact('post'));
+    public function edit(Post $post)
+    {
+        $this->authorize('update', $post);
+        return view('admin.posts.edit', compact('post'));
     }
-    public function postEdit(Post $post,Request $request) {
+    public function postEdit(Post $post, Request $request)
+    {
         $request->validate(
             [
                 'title' => 'required',
-                'content'=>'required',
+                'content' => 'required',
             ],
             [
                 'title.required' => 'Tên không được để trống',
@@ -62,11 +65,10 @@ class PostsController extends Controller
         $post->save();
         return back()->with('msg', 'Cập nhật bài viết thành công...');
     }
-    public function delete(Post $post){
-       
-            Post::destroy($post->id);
-            return redirect()->route('admin.posts.index')->with('msg', 'Xóa bài viết thành công');
-        
-        
+    public function delete(Post $post)
+    {
+        $this->authorize('delete', $post);
+        Post::destroy($post->id);
+        return redirect()->route('admin.posts.index')->with('msg', 'Xóa bài viết thành công');
     }
 }
